Move Microblink config into single class
