--- enrollment-server-onboarding-provider-microblink/src/main/java/com/wultra/app/onboardingserver/provider/microblink/MicroblinkRestClientConfig.java
@@ -1,63 +1,0 @@
-/*
- * PowerAuth Enrollment Server
- * Copyright (C) 2025 Wultra s.r.o.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.wultra.app.onboardingserver.provider.microblink;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.wultra.app.onboardingserver.provider.microblink.api.DocumentVerificationResponseParser;
-import com.wultra.core.rest.client.base.DefaultRestClient;
-import com.wultra.core.rest.client.base.RestClient;
-import com.wultra.core.rest.client.base.RestClientException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-
-/**
- * Configuration for Microblink {@link RestClient}.
- *
- * @author Michal Rozehnal, michal.rozehnal@wultra.com
- */
-@ConditionalOnProperty(value = "enrollment-server-onboarding.document-verification.provider", havingValue = "microblink")
-@Configuration
-@Slf4j
-class MicroblinkRestClientConfig {
-
-    @Bean("microblinkRestClient")
-    RestClient microblinkRestClient(MicroblinkConfigProperties properties) throws RestClientException {
-        logger.info("Registering Microblink RestClient with base URL: {}", properties.getRestClientConfig().getBaseUrl());
-
-        final var httpHeaders = new HttpHeaders();
-        httpHeaders.add(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE);
-
-        final var restClientConfig = properties.getRestClientConfig();
-        restClientConfig.setDefaultHttpHeaders(httpHeaders);
-
-        return new DefaultRestClient(restClientConfig);
-    }
-
-    @Bean("microblinkDocumentVerificationResponseParser")
-    DocumentVerificationResponseParser documentVerificationResponseParser() {
-        final var objectMapper = new ObjectMapper()
-                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        return new DocumentVerificationResponseParser(objectMapper);
-    }
-}
